+++ src/qei.rs
@@ -3,6 +3,7 @@ use crate::{
     hal::{self, Direction},
     pac::RCC,
     rcc,
+    timer::General,
 };
 
 pub trait Pins<TIM> {}
@@ -48,7 +49,7 @@ impl<TIM: Instance, PINS> Qei<TIM, PINS> {
 }
 
 impl<TIM: Instance, PINS> hal::Qei for Qei<TIM, PINS> {
-    type Count = TIM::Count;
+    type Count = TIM::Width;
 
     fn count(&self) -> Self::Count {
         self.tim.read_count() as Self::Count
@@ -63,23 +64,22 @@ impl<TIM: Instance, PINS> hal::Qei for Qei<TIM, PINS> {
     }
 }
 
-pub trait Instance: crate::Sealed + rcc::Enable + rcc::Reset {
-    type Count;
-
+pub trait Instance: crate::Sealed + rcc::Enable + rcc::Reset + General {
     fn setup_qei(&mut self);
-    fn read_count(&self) -> Self::Count;
+
     fn read_direction(&self) -> bool;
 }
 
 macro_rules! hal {
-    ($($TIM:ty: ($bits:ident),)+) => {
+    ($($TIM:ty,)+) => {
         $(
             impl Instance for $TIM {
-                type Count = $bits;
-
                 fn setup_qei(&mut self) {
                     // Configure TxC1 and TxC2 as captures
-                    self.ccmr1_output()
+                    #[cfg(not(feature = "stm32f410"))]
+                    self.ccmr1_input().write(|w| w.cc1s().ti1().cc2s().ti2());
+                    #[cfg(feature = "stm32f410")]
+                    self.ccmr1_input()
                         .write(|w| unsafe { w.cc1s().bits(0b01).cc2s().bits(0b01) });
                     // enable and configure to capture on rising edge
                     self.ccer.write(|w| {
@@ -92,17 +92,12 @@ macro_rules! hal {
                             .cc2p()
                             .clear_bit()
                     });
-                    // configure as quadrature encoder
-                    // some chip variants declare `.bits()` as unsafe, some don't
-                    #[allow(unused_unsafe)]
+                    #[cfg(not(feature = "stm32f410"))]
+                    self.smcr.write(|w| w.sms().encoder_mode_3());
+                    #[cfg(feature = "stm32f410")]
                     self.smcr.write(|w| unsafe { w.sms().bits(3) });
-                    #[allow(unused_unsafe)]
-                    self.arr.write(|w| unsafe { w.bits($bits::MAX as u32) });
+                    self.set_auto_reload(<$TIM as General>::Width::MAX as u32).unwrap();
                     self.cr1.write(|w| w.cen().set_bit());
-                }
-
-                fn read_count(&self) -> Self::Count {
-                    self.cnt.read().bits() as Self::Count
                 }
 
                 fn read_direction(&self) -> bool {
@@ -114,18 +109,18 @@ macro_rules! hal {
 }
 
 hal! {
-    crate::pac::TIM1: (u16),
-    crate::pac::TIM5: (u32),
+    crate::pac::TIM1,
+    crate::pac::TIM5,
 }
 
 #[cfg(feature = "tim2")]
 hal! {
-    crate::pac::TIM2: (u32),
-    crate::pac::TIM3: (u16),
-    crate::pac::TIM4: (u16),
+    crate::pac::TIM2,
+    crate::pac::TIM3,
+    crate::pac::TIM4,
 }
 
 #[cfg(feature = "tim8")]
 hal! {
-    crate::pac::TIM8: (u16),
+    crate::pac::TIM8,
 }